WIP Hyprland config
I used this to try out Hyprland. It might actually be too minimal
for me; there's a lot it lacks and I think I'd need to do quite a
bit more configuration to get it where I want to. I think I will
stick with KDE Plasma for now but perhaps revisit Hyprland in the
future.